+++ apps/backoffice-v2/src/pages/Root/Root.page.tsx
@@ -3,8 +3,6 @@ import { Providers } from '../../common/components/templates/Providers/Providers
 import { useCustomerQuery } from '@/domains/customer/hooks/queries/useCustomerQuery/useCustomerQuery';
 import { FullScreenLoader } from '@/common/components/molecules/FullScreenLoader/FullScreenLoader';
 import Chatbot from '@/domains/chat/chatbot-opengpt';
-import { RenderChildrenInIFrame } from '@/common/components/organisms/RenderChildrenInIFrame/RenderChildrenInIFrame';
-import { ctw } from '@/common/utils/ctw/ctw';
 import { env } from '@/common/env/env';
 import { Outlet } from 'react-router-dom';
 import { ServerDownLayout } from './ServerDown.layout';
@@ -35,18 +33,11 @@ const ChatbotLayout: FunctionComponent = () => {
   const botpressClientId = customer?.features?.chatbot?.clientId || env.VITE_BOTPRESS_CLIENT_ID;
 
   return (
-    <RenderChildrenInIFrame
-      className={ctw('fixed bottom-right-0', {
-        'h-[700px] w-[400px]': isWebchatOpen,
-        'd-[80px]': !isWebchatOpen,
-      })}
-    >
-      <Chatbot
-        isWebchatOpen={isWebchatOpen}
-        toggleIsWebchatOpen={toggleIsWebchatOpen}
-        botpressClientId={botpressClientId}
-      />
-    </RenderChildrenInIFrame>
+    <Chatbot
+      isWebchatOpen={isWebchatOpen}
+      toggleIsWebchatOpen={toggleIsWebchatOpen}
+      botpressClientId={botpressClientId}
+    />
   );
 };