+++ apps/workflows-dashboard/src/pages/Filters/types/filters-query-params.ts
@@ -1,4 +1,5 @@
 export interface FiltersPageFilterQuery {
+  projectId: string;
   page?: number;
   limit?: number;
 }