+++ apps/workflows-dashboard/src/pages/WorkflowDefinitions/components/molecules/WorkflowDefinitionsTable/WorkflowDefinitionsTable.tsx
@@ -117,3 +117,5 @@ export const WorkflowDefinitionsTable = memo(({ items, isFetching, sorting, onSo
     </div>
   );
 });
+
+WorkflowDefinitionsTable.displayName = 'WorkflowDefinitionsTable';