+++ packages/rules-engine/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ballerine/rules-engine-lib
 
+## 0.5.27
+
+### Patch Changes
+
+- Bump
+
 ## 0.5.26
 
 ### Patch Changes